refactor lesson 1 out
