+++ src/RobotSimulator/ChallengeConfigLoader.ts
@@ -1,21 +1,15 @@
-import { ArenaConfig, getArenaConfig } from "./ArenaConfigLoader";
-import { CoreSpecs, CoreSimTypes } from "@fruk/simulator-core";
+import { getArenaConfig } from "./ArenaConfigLoader";
+import { ChallengeConfig } from "./Areanas/base";
+import { createFinishZoneSpec, expand } from "./Areanas/common";
+import * as Lesson1 from "./Areanas/lesson1";
 
 let challengeConfigs: Array<ChallengeConfig> = [
-  setupLesson1ChallengeA(),
-  setupLesson1ChallengeB(),
+  ...expand(Lesson1.challenges),
   setupParkingLotChallenge1(),
   setupParkingLotChallenge2(),
   setupZigZagChallenge(),
   setupBowlingChallenge(),
 ];
-
-export interface ChallengeConfig {
-  name: string;
-  startPosition: CoreSimTypes.Vector2d;
-  finishZoneSpec?: CoreSpecs.IZoneSpec;
-  arenaConfig: ArenaConfig;
-}
 
 export function getChallengesPerArena(): Map<string, Array<string>> {
   let challengesPerArena = new Map();
@@ -52,29 +46,7 @@ export function getChallengeConfig(name: string): ChallengeConfig {
   }
 
   // return default if 'name' not found
-  return setupLesson1ChallengeA();
-}
-
-function setupLesson1ChallengeA(): ChallengeConfig {
-  const challengeConfig: ChallengeConfig = {
-    name: "Lesson 1 - Challenge A",
-    startPosition: { x: 0, y: 8 },
-    finishZoneSpec: createFinishZoneSpec({ x: 0, y: -8 }),
-    arenaConfig: getArenaConfig("Lesson 1"),
-  };
-
-  return challengeConfig;
-}
-
-function setupLesson1ChallengeB(): ChallengeConfig {
-  const challengeConfig: ChallengeConfig = {
-    name: "Lesson 1 - Challenge B",
-    startPosition: { x: 8, y: 8 },
-    finishZoneSpec: createFinishZoneSpec({ x: -8, y: 8 }),
-    arenaConfig: getArenaConfig("Lesson 1"),
-  };
-
-  return challengeConfig;
+  return Lesson1.challenges[0]();
 }
 
 function setupParkingLotChallenge1(): ChallengeConfig {
@@ -118,19 +90,4 @@ function setupBowlingChallenge(): ChallengeConfig {
   };
 
   return challengeConfig;
-}
-
-function createFinishZoneSpec(
-  initialPosition: CoreSimTypes.Vector2d
-): CoreSpecs.IZoneSpec {
-  let finishZoneSpec: CoreSpecs.IZoneSpec = {
-    type: "zone",
-    zoneId: "finish",
-    xLength: 2,
-    zLength: 2,
-    baseColor: 0x00ff00,
-    initialPosition: initialPosition,
-  };
-
-  return finishZoneSpec;
 }
